chore: split macro test cases into separate files
